+++ examples/complete/tamr_vm.tf
@@ -17,6 +17,21 @@ data "aws_ami" "tamr-vm" {
   }
 }
 
+data "template_file" "install_nginx" {
+  template = file("${path.module}/files/install-nginx.tpl")
+  vars = {
+    tamr_unify_port = var.tamr_unify_port
+  }
+}
+
+data "template_file" "setup_dms" {
+  template = file("${path.module}/files/setup-dms.tpl")
+  vars = {
+    tamr_unify_port = var.tamr_unify_port
+    tamr_dms_port   = var.tamr_dms_port
+  }
+}
+
 module "sg_vm_web" {
   source = "git::git@github.com:Datatamer/terraform-aws-security-groups.git?ref=1.0.0"
   vpc_id = module.tamr_networking.vpc_id
@@ -26,25 +41,24 @@ module "sg_vm_web" {
   egress_protocol         = "all"
   ingress_security_groups = [module.alb.lb_security_group_id]
   ingress_protocol        = "tcp"
-  ingress_ports           = [var.tamr_unify_port]
-  sg_name_prefix          = var.name-prefix
+  ingress_ports           = [var.tamr_unify_port, var.tamr_dms_port]
+  sg_name_prefix          = format("%s-%s", "example-complete", "tamr-vm")
+
 }
 
 module "tamr-vm" {
   source                      = "git::git@github.com:Datatamer/terraform-aws-tamr-vm.git?ref=3.2.0"
-  aws_role_name               = format("%s-tamr-ec2-role", var.name-prefix)
-  aws_instance_profile_name   = format("%s-tamr-ec2-instance-profile", var.name-prefix)
-  aws_emr_creator_policy_name = format("%sEmrCreatorPolicy", var.name-prefix)
+  aws_role_name               = format("%s-tamr-ec2-role", var.name_prefix)
+  aws_instance_profile_name   = format("%s-tamr-ec2-instance-profile", var.name_prefix)
+  aws_emr_creator_policy_name = format("%sEmrCreatorPolicy", var.name_prefix)
   ami                         = local.ami_id
   instance_type               = "r5.2xlarge"
   key_name                    = var.key_pair
   availability_zone           = local.az
   vpc_id                      = module.tamr_networking.vpc_id
   subnet_id                   = module.tamr_networking.application_subnet_id
-  bootstrap_scripts = [
-    file("./install-nginx.sh")
-  ]
-  s3_policy_arns = []
+  bootstrap_scripts           = [data.template_file.setup_dms.rendered]
+  s3_policy_arns              = []
   depends_on = [
     module.tamr_networking
   ]