+++ examples/complete/variables.tf
@@ -1,7 +1,11 @@
 variable "ingress_cidr_blocks" {
   type        = list(string)
   description = "The cidr range that will be accessing the tamr vm"
-  default     = ["0.0.0.0/0"]
+}
+
+variable "egress_cidr_blocks" {
+  type        = list(string)
+  description = "The cidr ranges that will be accessible from EMR"
 }
 
 variable "tls_certificate_arn" {
@@ -20,13 +24,25 @@ variable "tamr_unify_port" {
   default     = "9100"
 }
 
+variable "tamr_dms_port" {
+  type        = string
+  description = "Identifies the DMS access HTTP port"
+  default     = "9155"
+}
+
 variable "ami_id" {
   type        = string
   description = "The AMI to use for the tamr vm"
   default     = ""
 }
 
-variable "name-prefix" {
+variable "abac_valid_tags" {
+  type        = map(list(string))
+  description = "Valid tags for maintaining resources when using ABAC IAM Policies with Tag Conditions. Make sure `tags` contain a key value specified here."
+  default     = {}
+}
+
+variable "name_prefix" {
   type        = string
   description = ""
   default     = "tamr-"
@@ -36,11 +52,21 @@ variable "key_pair" {
   type = string
 }
 
+variable "bucket_name_for_logs" {
+  type        = string
+  description = "S3 bucket name for cluster logs."
+}
+
+variable "bucket_name_for_root_directory" {
+  type        = string
+  description = "S3 bucket name for storing root directory"
+}
+
 variable "tags" {
   type        = map(string)
   description = "A map of tags to add to all resources."
   default = {
-    "Name" : "tamr-vpc"
+    "Name" : "tamr-vpc",
     "application" : "tamr",
     "Terraform" : "true"
   }