+++ examples/complete/vpc.tf
@@ -1,4 +1,5 @@
 module "tamr_networking" {
+
   source                             = "../../"
   vpc_cidr_block                     = "10.0.0.0/16"
   application_subnet_cidr_block      = "10.0.0.0/24"
@@ -9,7 +10,9 @@ module "tamr_networking" {
   ingress_cidr_blocks                = var.ingress_cidr_blocks
   availability_zones                 = var.availability_zones
   create_public_subnets              = true
-  enable_nat_gateway                 = true
+  enable_nat_gateway                 = false
   create_load_balancing_subnets      = true
+  name_prefix                        = var.name_prefix
+  interface_endpoint_ingress_sg      = module.sg_vm_web.security_group_ids[0]
   tags                               = var.tags
 }