+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Tamr Terraform AWS Networking Repo
+## v1.0.0 - November 29th 2021
+* Refactors ALB Module
+* Adds EMR to complete example
+* Adds Host-Based Routing
+* Adds interface endpoint blocks to allow Tamr VM and EMR to communicate without traversing the internet
+* Updates minimal and complete examples to allow the use of interface endpoint
+
 ## v0.1.0 - July 20th 2021
-* Added minimal VPC, NACLs, S3 VPC Endpoint and optional public subnets and NAT gateway.
+* Added minimal VPC, NACLs, S3 VPC Endpoint and optional public subnets and NAT gateway
 * Added example with ALB and optional TLS Certificate